+++ files/src/sdk/models/operations/deletefilestemplateid.ts
@@ -31,7 +31,7 @@ export class DeleteFilesTemplateIdResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Document Template