+++ files/src/sdk/models/operations/getfilesmemberpublicid.ts
@@ -37,5 +37,5 @@ export class GetFilesMemberPublicIdResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }