+++ files/src/sdk/models/operations/getfilestemplate.ts
@@ -23,7 +23,7 @@ export class GetFilesTemplateResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Describes a list of templates